+++ nmea_gps.h
@@ -17,10 +17,10 @@ extern int time_h ;// hh,mm,ss.ss
 extern int time_m ;
 extern int time_s ;
 extern int time_ms ;
-extern float base_latt;
-extern float base_long;
+extern double base_latt;
+extern double base_long;
 
-extern float magic_number;
+extern double magic_number;
 
 std::string build_nmea_msg(float x, float y, float z);
 // this number is magic and converts change in distance from m to deg (aprox)
